Refactor user ID type from Integer to Long throughout code

+++ springboot-mall/src/main/java/org/kevinlin/springbootmall/controller/UserController.java
@@ -21,7 +21,7 @@ public class UserController {
     @PostMapping("/users/register")
     public ResponseEntity<User> register(@RequestBody @Valid UserRegisterRequest userRegisterRequest) {
 
-        Integer userId = userService.register(userRegisterRequest);
+        Long userId = userService.register(userRegisterRequest);
 
         User user = userService.getUserById(userId);
