+++ springboot-mall/src/main/java/org/kevinlin/springbootmall/rowmapper/UserRowMapper.java
@@ -5,7 +5,6 @@ import org.springframework.jdbc.core.RowMapper;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.util.Date;
 
 public class UserRowMapper implements RowMapper<User> {
 
@@ -13,7 +12,7 @@ public class UserRowMapper implements RowMapper<User> {
     public User mapRow(ResultSet resultSet, int i) throws SQLException {
         User user = new User();
 
-        user.setUserId(resultSet.getInt("user_id"));
+        user.setUserId(resultSet.getLong("user_id"));
         user.setEmail(resultSet.getString("email"));
         user.setPassword(resultSet.getString("password"));
         user.setCreatedDate(resultSet.getTimestamp("created_date"));
